backend/auth.py
import os
import requests
from jose import jwt
from dotenv import load_dotenv

# fetch data from .env
load_dotenv()

SUPABASE_PROJECT_ID = os.getenv("SUPABASE_PROJECT_ID")
JWKS_URL = f"https://{SUPABASE_PROJECT_ID}.supabase.co/auth/v1/keys"

